dom/xul/test/test_bug583948.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

<body xmlns="http://www.w3.org/1999/xhtml">
  <div id="content" style="display: none"/>
</body>

<script>
SimpleTest.waitForExplicitFinish();

var attempts = 0;

function update() {
  // without the crash fix, this usually crashes after 2 to 4 reloads
  if (++attempts == 6) {
    ok(true, "didn't crash after 6 attempts");
    otherWindow.close();
    SimpleTest.waitForFocus(function() {
      SimpleTest.finish();
    });
  } else {
    otherWindow.document.commandDispatcher.updateCommands('');
    setTimeout(function() {
      otherWindow.location.reload()
    }, 0);
  }
}

var otherWindow = window.open("window_bug583948.xul", "_new", "chrome");
</script>

</window>
